주문 완료 로직, 주문 등록 뷰 개선

+++ src/main/java/com/example/zeronine/order/form/OrderForm.java
@@ -29,7 +29,7 @@ public class OrderForm {
     private String name;
 
     @NotNull
-    private Long numberOfLimit;
+    private Integer numberOfLimit;
 
     @NotNull
     private Long price;
